Refactor: replace components
